Rename and update BlackCircleWithPlusButton. Update
