+++ JourneysSquad/JourneysSquad/Helpers/AppStyle/String+AppIconName.swift
@@ -1,5 +1,10 @@
 extension String {
+    /// The set of names for using UI elements such as buttons, systemImages.
     enum AppIconName {
+        /// The name used to create an image with a plus sign.
         static let plus = "plus"
+
+        /// The text for `Save` button label.
+        static let save = "Save"
     }
 }
